isInert() allows the user to avoid creating children
One can implement isInert() in a child class that inherits
AdmissibilityCondition to specify a condition which will have the effect
of not creating a child (so use with caution!).

It is used when you know that a block is null and will always be null,
even after factorization, as it happens sometimes in sparse techniques
(nested dissection, symbolic factorization...).